move expand_message_xmd into a new module
